+++ R/utils-rd.R
@@ -112,12 +112,5 @@ make_as_character_rd <- function() {
 }
 
 has_topic <- function(topic, package) {
-  tryCatch(
-    {
-      out <- exec("help", topic, package, .env = global_env())
-      inherits(out, "dev_topic") ||
-        (inherits(out, "help_files_with_topic") && length(out) == 1)
-    },
-    error = function(c) FALSE
-  )
+  topic %in% pkg_topics(package)
 }